requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
openai==0.28.1
gunicorn==21.2.0
Werkzeug==2.3.7
